specs/controller show action (#57)
* Update drive_synchronizer.rb

bringing back .from_service_account_key(StringIO.new(@api_key))
deleted the variable @folder_name because we don't use it - if you want to work to refactor the folder_name let's please look into this together when I'm in the office. I've already tried and it was breaking the app. I think it's better if we memoize it (even tho, that was also breaking the app). Let's leave the code as it is for @eviltrout to check it and get some feedback from him.

* Delete LICENSE

* MIT license

we added the MIT license to our project

* Code of Conduct (#53)

a little extra morality

* add listener

* add listener (#54)

* specs for show action