controller specs (#58)
* Update drive_synchronizer.rb

bringing back .from_service_account_key(StringIO.new(@api_key))
deleted the variable @folder_name because we don't use it - if you want to work to refactor the folder_name let's please look into this together when I'm in the office. I've already tried and it was breaking the app. I think it's better if we memoize it (even tho, that was also breaking the app). Let's leave the code as it is for @eviltrout to check it and get some feedback from him.

* Delete LICENSE

* MIT license

we added the MIT license to our project

* Code of Conduct (#53)

a little extra morality

* add listener

* first doodle structure of a downloader

* add to_json

* download structure

* attr

* pick file

* create url method

* assign id

* taking the memoized google_files

* job of sending an email

* job listens to an event of choosing a file

* load the downloader files in the plugin.rb

* making the job work

* create JSON object

* fix naming & create JSON obj

* fix naming

* complete JSON obj

* add JSON puts

* created_at in json

* added controller / draw route

* display JSON in browser

* use nil to initialise DriveDownloader

* change route to match google plugin

* trying to build a put/post for the job

* email action

* trying to pass the id

* try to get id

* specs

* specs for downloaders_controller

* controller specs

* naming

* right tests